Change .env file handling and file storage
The .env file has been updated to support two ways of providing a measurement file storage directory:
- give a relative path using VITE_BACKEND_MEASUREMENT_DIR (as before) and have the API auto-detect the preferred location (LOCALAPPDATA on Windows; XDG data directory on Linux)
- give a full absolute path that is unchecked and to be used at your own discretion via the variable VITE_BACKEND_FULL_MEASUREMENT_PATH

We ought to consider removing the .env file from git tracking as it was only added to provide ease-of-use when starting to work with the project. Now, the ability to run different settings on different systems is required and we do not want to commit those settings.