Rework Sensor & Holder System
This makes the API the source of truth for sensors and holders in the system.
The sensors.yaml file holds all sensors and configurations ("holders") available.

It needs to be either in the project root (for running as-is) or in the bundle dir
(running as api.exe in an installer state).

The reset sensor endpoint was removed as it is not sensible given the API is the SoT.