Rework file storage settings
This changes how the file structure works. We now store everything that
we need for normal operation in the user_data_dir and ONLY use the
bundle dir if we detect that the config directory does not have the same
files present.

This also changes that measurement files are now under a new path.